+++ src/main/kotlin/picklab/backend/common/model/BusinessException.kt
@@ -1,0 +1,5 @@
+package picklab.backend.common.model
+
+class BusinessException(
+    val errorCode: ErrorCode,
+) : RuntimeException(errorCode.message)
