--- src/main/kotlin/picklab/backend/member/domain/MemberRepository.kt
@@ -1,6 +1,0 @@
-package picklab.backend.member.domain
-
-import org.springframework.data.jpa.repository.JpaRepository
-import picklab.backend.member.domain.entity.Member
-
-interface MemberRepository : JpaRepository<Member, Long>
