+++ src/main/kotlin/picklab/backend/member/domain/repository/InterestedJobCategoryRepository.kt
@@ -1,4 +1,4 @@
-package picklab.backend.member.domain
+package picklab.backend.member.domain.repository
 
 import org.springframework.data.jpa.repository.JpaRepository
 import picklab.backend.member.domain.entity.InterestedJobCategory
